Update README.md
Mental health is creating chaos all around the world. Thankfully, people are noticing this issue and giving importance to it. This is the reason why demand for VigorVita CBD Gummies has increased rapidly. Now, people know that long-term stress and anxiety are not normal and that they need to be treated at the right time. Some people ignore it, and with leading days they start suffering from high blood pressure, schizophrenia, cardiovascular health issues, diabetes, and even scientific research has proven that one of the major reasons for cancer is stress and pressure that a person cannot handle, but the person has been suffering from that stress and pressure for a long time. If you remain stress-free, then only you can be productive or enjoy your life.